Move `CustomRuleTimer` to its own file
And add docs.
