+++ Source/SwiftLintFramework/Rules/Style/CustomRules.swift
@@ -1,29 +1,5 @@
 import Foundation
 
-public final class CustomRuleTimer {
-    private let lock = NSLock()
-    private var ruleIDForTimes = [String: [TimeInterval]]()
-    fileprivate var shouldRecord = false
-
-    public static let shared = CustomRuleTimer()
-
-    public func activate() {
-        shouldRecord = true
-    }
-
-    func register(time: TimeInterval, forRuleID ruleID: String) {
-        guard shouldRecord else { return }
-
-        lock.lock()
-        defer { lock.unlock() }
-        ruleIDForTimes[ruleID, default: []].append(time)
-    }
-
-    public func dump() -> [String: TimeInterval] {
-        ruleIDForTimes.mapValues { $0.reduce(0, +) }
-    }
-}
-
 private extension Region {
     func isRuleDisabled(customRuleIdentifier: String) -> Bool {
         return disabledRuleIdentifiers.contains(RuleIdentifier(customRuleIdentifier))
